+++ app/features/__init__.py
@@ -1,0 +1,3 @@
+from app.features import invites, issues, vouches
+
+__all__ = ("invites", "issues", "vouches")
